backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.32
greenlet==3.0.3
asyncpg>=0.29.0,<0.31.0
alembic==1.13.2
redis==5.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0
httpx==0.27.0
pytest==8.3.2
pytest-asyncio==0.23.8
asyncpg==0.30.0
